requirements.txt
boto3==1.28.57
botocore==1.31.57
humanize==4.8.0
jmespath==1.0.1
python-dateutil==2.8.2
python-decouple==3.8
s3transfer==0.7.0
six==1.16.0
urllib3==1.26.16
